+++ src/Serializer/IdNormalizer.php
@@ -18,7 +18,7 @@ final class IdNormalizer implements NormalizerInterface, DenormalizerInterface, 
     }
 
     /**
-     * @param class-string                      $type
+     * @param string                            $type
      * @param array<string, string|int|boolean> $context
      */
     public function supportsDenormalization($data, $type, $format = null, array $context = []): bool